pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cross-sectional-machine-learning"
version = "0.1.0"
description = "Cross-sectional factor mining with pluggable machine learning models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=1.23",
  "pandas>=2.0",
  "pandas-ta>=0.3.14b0",
  "pyarrow>=12.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "scikit-learn>=1.2",
  "tushare>=1.2.89",
  "xgboost>=1.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov",
]
rqdata = [
  "rqdatac>=3.4.3",
  "rqdatac-hk",
  "rqsdk",
]
stats = [
  "scipy>=1.10",
]

[project.scripts]
csml = "csml.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
csml = ["config/*.yml"]

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1 --cov=csml --cov-report=term-missing"
markers = [
  "integration: end-to-end or slower tests",
  "slow: intentionally slow tests",
]
